Split companies transformers into seperate files
