+++ src/apps/companies/transformers/company-response-to-one-list-view-record.js
@@ -1,0 +1,14 @@
+/* eslint-disable camelcase */
+const { get } = require('lodash')
+
+const { getDataLabels } = require('../../../lib/controller-utils')
+const { accountManagementDisplayLabels } = require('../labels')
+
+module.exports = function transformCompanyResponseToOneListViewRecord ({ one_list_account_owner, classification }) {
+  const viewRecord = {
+    oneListAccountManager: get(one_list_account_owner, 'name', 'None'),
+    oneListTier: get(classification, 'name', 'None'),
+  }
+
+  return getDataLabels(viewRecord, accountManagementDisplayLabels)
+}
