+++ src/apps/companies/transformers/index.js
@@ -1,0 +1,15 @@
+const transformCompaniesHouseResponseToListItem = require('./companies-house-response-to-list-item')
+const transformCompaniesHouseResponseToViewRecord = require('./companies-house-response-to-view-record')
+const transformCompanyResponseToForm = require('./company-response-to-form')
+const transformCompanyResponseToListItem = require('./company-response-to-list-item')
+const transformCompanyResponseToOneListViewRecord = require('./company-response-to-one-list-view-record')
+const transformCompanyResponseToViewRecord = require('./company-response-to-view-record')
+
+module.exports = {
+  transformCompaniesHouseResponseToListItem,
+  transformCompaniesHouseResponseToViewRecord,
+  transformCompanyResponseToForm,
+  transformCompanyResponseToListItem,
+  transformCompanyResponseToOneListViewRecord,
+  transformCompanyResponseToViewRecord,
+}
